set up radar chart component to get an array of audio features for user's top 15 tracks
